Generic javax.servlet.ServletContext interceptor
The patch tries to intercept any class implementing javax.servlet.ServletContext
with notable exception of Jetty (due to apparent initialization issues that are handled
better in a separate interceptor) and Tomcat's ApplicationContextFacade (it's enough
to just capture the ApplicationContext).

Tested with Weblogic 12.2.1.3 (unfortunately, due to complexity of providing custom JAR
for the prioprietary software no specific unit test provided)
