+++ rules/opentracing-specialagent-web-servlet-filter/src/main/java/io/opentracing/contrib/specialagent/webservletfilter/ServletContextAgentIntercept.java
@@ -17,16 +17,24 @@ package io.opentracing.contrib.specialagent.webservletfilter;
 
 import java.util.EnumSet;
 
-import org.apache.catalina.core.ApplicationContext;
-
 import io.opentracing.contrib.web.servlet.filter.TracingFilter;
 import io.opentracing.util.GlobalTracer;
 
-public class TomcatAgentIntercept {
+import javax.servlet.FilterRegistration;
+import javax.servlet.ServletContext;
+
+public class ServletContextAgentIntercept {
   public static void exit(final Object thiz) {
-    final ApplicationContext context = (ApplicationContext)thiz;
-    final TracingFilter filter = new TracingFilter(GlobalTracer.get());
-    final String[] patterns = {"/*"};
-    context.addFilter("tracingFilter", filter).addMappingForUrlPatterns(EnumSet.allOf(javax.servlet.DispatcherType.class), true, patterns);
+    if (thiz instanceof ServletContext) {
+      final ServletContext context = (ServletContext) thiz;
+
+      final TracingFilter filter = new TracingFilter(GlobalTracer.get());
+      final String[] patterns = {"/*"};
+
+      final FilterRegistration.Dynamic registration = context.addFilter("tracingFilter", filter);
+      if (registration != null) {
+        registration.addMappingForUrlPatterns(EnumSet.allOf(javax.servlet.DispatcherType.class), true, patterns);
+      }
+    }
   }
 }
